Python Span linked to C++ Span
The Python span is now attached to the Span in C++ when the
Proxy span is created from Python. And optionally empty when
created purely in C++. The plan is to attach it when it is
first created in Python during the observation step.

A PythonReference wrapper is used for keeping references to
python objects managed in C++ shared_ptr, even though they
will only ever by accessed within Cython.